bigger integer types should enable the usage of this data struct for race results

+++ include/core/data/internal/LapTime.h
@@ -12,7 +12,7 @@ namespace Lap::Internal {
 
         Time() = default;
         Time(const uint32_t milliseconds);
-        Time(const uint8_t seconds, const uint16_t milliseconds);
+        Time(const uint32_t seconds, const uint16_t milliseconds);
         Time(const Time& other) = default;
         Time& operator=(const Time& other);
         Time& operator+=(const Time& other);
@@ -29,7 +29,7 @@ namespace Lap::Internal {
         const std::string formattedPrint(bool minutes = false) const;
         const bool valid() const;
 
-        uint8_t m_seconds;
+        uint32_t m_seconds;
         uint16_t m_milliseconds;
 
     };
